修复删除水印模板失败的问题。

+++ src/renderer/pages/watermark.vue
@@ -446,7 +446,9 @@ export default {
             }
           }
           let template = this.$store.state.watermark.templates[index]
-          fs.unlinkSync(template.image)
+          if (template.image) {
+            fs.unlinkSync(template.image)
+          }
           this.$store.dispatch('watermark/templateDelete', index)
         }
       })
